sdk/requirements.txt
# CosmoEmbeddings SDK Requirements

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
sentence-transformers>=2.2.0
pynacl>=1.5.0
skyfield>=1.42
requests>=2.26.0
python-dotenv>=0.19.0
tqdm>=4.62.0
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.9b0
flake8>=3.9.2
mypy>=0.910

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0

# Testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Development
black>=21.5b2
flake8>=3.9.2
mypy>=0.910

# Embedding models
sentence-transformers>=2.2.2

# Requests
requests>=2.26.0 